digital_campus/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.35.87
botocore==1.35.87
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
Collectfast==2.2.0
constantly==23.10.4
crispy-bootstrap4==2024.10
cryptography==44.0.0
daphne==4.1.2
Django==5.2.1
django-autocomplete-light==3.11.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-storages==1.14.4
django-taggit==6.1.0
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
joblib==1.4.2
msgpack==1.1.0
numpy==2.0.2
Pillow==9.5.0
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
pytz==2024.2
RapidFuzz==3.11.0
redis==5.2.1
s3transfer==0.10.4
scikit-learn==1.6.1
scipy==1.13.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
svgwrite==1.4.3
threadpoolctl==3.6.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==1.26.20
websocat==1.13.0
whitenoise==6.8.2
zope.interface==7.2
